Move file upload functionality from the form component to the controller
