+++ app/routes/add-preprint.js
@@ -2,7 +2,4 @@ import Ember from 'ember';
 import attr from 'ember-data/attr';
 
 export default Ember.Route.extend({
-    model() {
-        //return this.get('store').findAll('preprint');
-    }
-});
+    });
