+++ app/templates/add-preprint.hbs
@@ -29,7 +29,7 @@
 
         </div> <!--END COL 7 DIV-->
         <div class="col-lg-5"> <!--COL 5 DIV-->
-            {{preprint-form uploadFile=(action 'uploadPreprintFile')}}
+            {{preprint-form buttonAction=(action 'uploadNewPreprintToNewProject')}}
         </div> <!--END COL 5 DIV-->
     </div> <!--ROW DIV-->
     <div class="row"> <!--2ND ROW DIV-->
