--- tests/unit/controllers/add-preprint-test.js
@@ -1,12 +1,0 @@
-import { moduleFor, test } from 'ember-qunit';
-
-moduleFor('controller:add-preprint', 'Unit | Controller | add preprint', {
-  // Specify the other units that are required for this test.
-  // needs: ['controller:foo']
-});
-
-// Replace this with your real tests.
-test('it exists', function(assert) {
-  let controller = this.subject();
-  assert.ok(controller);
-});
